fix: renames 'FieldContstraints' to 'FieldConstraints' (#2352)
